cleanup rhizomeutils
